Moved multiboot files into architecture They are x86 family only, so they belong there, but I don't know where.
